the core package isn't really carrying its weight quite yet so I'm moving this in with the rest of the common analytics classes.
