+++ google-cloud-tools-plugin/google-cloud-core/src/com/google/cloud/tools/intellij/GoogleCloudCoreMessageBundle.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 Google Inc. All Rights Reserved.
+ * Copyright 2018 Google Inc. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.intellij.core;
+package com.google.cloud.tools.intellij;
 
 import com.intellij.CommonBundle;
 import java.util.ResourceBundle;
