+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/debugger/CloudDebugProcessWatcher.java
@@ -16,8 +16,8 @@
 
 package com.google.cloud.tools.intellij.debugger;
 
+import com.google.cloud.tools.intellij.analytics.GctTracking;
 import com.google.cloud.tools.intellij.analytics.UsageTrackerProvider;
-import com.google.cloud.tools.intellij.core.analytics.GctTracking;
 import com.google.cloud.tools.intellij.util.GctBundle;
 import com.intellij.execution.Executor;
 import com.intellij.execution.ProgramRunnerUtil;
