+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/resources/ResourceSelectorLoadingItem.java
@@ -16,7 +16,7 @@
 
 package com.google.cloud.tools.intellij.resources;
 
-import com.google.cloud.tools.intellij.core.GoogleCloudCoreIcons;
+import com.google.cloud.tools.intellij.GoogleCloudCoreIcons;
 import com.intellij.ui.components.JBLabel;
 import java.awt.Color;
 import java.awt.FlowLayout;
