+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/ui/DisablePluginWarningDialog.java
@@ -16,8 +16,8 @@
 
 package com.google.cloud.tools.intellij.ui;
 
+import com.google.cloud.tools.intellij.analytics.GctTracking;
 import com.google.cloud.tools.intellij.analytics.UsageTrackerProvider;
-import com.google.cloud.tools.intellij.core.analytics.GctTracking;
 import com.google.cloud.tools.intellij.util.GctBundle;
 import com.intellij.ide.plugins.IdeaPluginDescriptor;
 import com.intellij.ide.plugins.PluginManager;
